requirements.txt
ffmpeg-python
nltk
openai-whisper
pydub
python-docx
streamlit
streamlit-audiorecorder
openai
groq
pydub
markdown
pypandoc
requests